% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculatePIscore.R
\name{adjustExpressionToCD45}
\alias{adjustExpressionToCD45}
\title{Adjusting expression to CD45}
\usage{
adjustExpressionToCD45(expr, fit_params)
}
\arguments{
\item{expr}{the gene expression data set. A matrix with row names as symbols and columns as samples.}

\item{fit_params}{a vector of the curves of association between genes and CD45 learned from GTEx.}
}
\value{
the adjusted expression profile
}
\description{
\code{adjustExpressionToCD45} Returns the adjusted expression profile.
}
