% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculatePIscore.R
\name{calculatePIscore}
\alias{calculatePIscore}
\title{Calculating the PI score}
\usage{
calculatePIscore(expr, tissue)
}
\arguments{
\item{expr}{the gene expression data set. A matrix with row names as symbols and columns as samples.}

\item{tissue}{the tissue source of the expression profile. Must be from the list of tissues in fit_params$tissues}

\item{fit_params}{the cruves of association with CD45 learned from GTEx.}
}
\value{
the PI scores
}
\description{
\code{calculatePIscore} Returns the PI scores based on expression profiles.
}
