% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculatePIscore.R
\name{createParams}
\alias{createParams}
\title{Calculating the parameters}
\usage{
createParams()
}
\value{
the parameters matrix
}
\description{
\code{createParams} Returns a matrix of the parameters learned from GTEx.
}
