% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculatePIscore.R
\docType{data}
\name{fit_params}
\alias{fit_params}
\title{CD45 curves learned from GTEx}
\format{list:
\describe{
  \item{coeff}{a matrix with CD45 curves learned from GTEx}
  \item{tissues}{tissue names}
}}
\usage{
fit_params
}
\description{
CD45 curves learned from GTEx
}
\keyword{datasets}
