% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculatePIscore.R
\docType{package}
\name{Parainflammation}
\alias{Parainflammation}
\alias{Parainflammation-package}
\title{Parainflammation: A package for calculating the parainflammation score.}
\description{
Parainflammation: A package for calculating the parainflammation score.
}
