% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculatePIscore.R
\docType{data}
\name{fit_params}
\alias{fit_params}
\title{Parainflammation datasets}
\format{list:
\describe{
  \item{fit_params}{CD45 curves learned from GTEx}
}}
\usage{
fit_params
}
\description{
Parainflammation datasets
}
\keyword{datasets}
